requirements.txt
selenium>=4.21.0
Flask>=3.0.3
python-dotenv>=1.0.1
Flask-HTTPAuth>=4.8.0